--- resources/views/exercisegps/styles.css
@@ -1,12 +1,0 @@
-#lblStatus {
-  x: 8;
-  y: 100%-18;
-  text-length: 20;
-}
-
-#boxStats {
-  x: 0;
-  y: 0;
-  width: 100%;
-  height: 100%;
-}
